updated and tested state changes requested
